+++ compute/cloud-client/src/test/java/compute/reservation/ReservationIT.java
@@ -19,6 +19,7 @@ package compute.reservation;
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
 import static compute.Util.getZone;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import com.google.api.gax.rpc.NotFoundException;
 import com.google.cloud.compute.v1.AttachedDisk;
@@ -38,11 +39,13 @@ import compute.Util;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
+import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import org.junit.Assert;
+import org.junit.FixMethodOrder;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -52,14 +55,17 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
+import org.junit.runners.MethodSorters;
 
 @RunWith(JUnit4.class)
 @Timeout(value = 10, unit = TimeUnit.MINUTES)
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class ReservationIT {
 
   private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
   private static final String ZONE = getZone();
   private static ReservationsClient reservationsClient;
+  private static String RESERVATION_NAME;
   private static String RESERVATION_NAME_GLOBAL;
   private static String RESERVATION_NAME_REGIONAL;
   private static String GLOBAL_INSTANCE_TEMPLATE_URI;
@@ -92,9 +98,9 @@ public class ReservationIT {
     // Cleanup existing stale resources.
     Util.cleanUpExistingInstanceTemplates("test-global-inst-temp", PROJECT_ID);
     Util.cleanUpExistingRegionalInstanceTemplates("test-regional-inst-temp", PROJECT_ID, ZONE);
-    Util.cleanUpExistingReservations("test-reserv-regional", PROJECT_ID, ZONE);
-    Util.cleanUpExistingReservations("test-reserv-global", PROJECT_ID, ZONE);
+    Util.cleanUpExistingReservations("test-reserv-", PROJECT_ID, ZONE);
 
+    RESERVATION_NAME = "test-reserv-" + UUID.randomUUID();
     RESERVATION_NAME_GLOBAL = "test-reserv-global-" + UUID.randomUUID();
     RESERVATION_NAME_REGIONAL = "test-reserv-regional-" + UUID.randomUUID();
     GLOBAL_INSTANCE_TEMPLATE_URI = String.format("projects/%s/global/instanceTemplates/%s",
@@ -138,11 +144,10 @@ public class ReservationIT {
             + REGIONAL_INSTANCE_TEMPLATE_NAME);
 
     // Delete all reservations created for testing.
+    DeleteReservation.deleteReservation(PROJECT_ID, ZONE, RESERVATION_NAME);
     DeleteReservation.deleteReservation(PROJECT_ID, ZONE, RESERVATION_NAME_GLOBAL);
     DeleteReservation.deleteReservation(PROJECT_ID, ZONE, RESERVATION_NAME_REGIONAL);
 
-    assertThat(stdOut.toString()).contains("Deleted reservation: " + RESERVATION_NAME_GLOBAL);
-    assertThat(stdOut.toString()).contains("Deleted reservation: " + RESERVATION_NAME_REGIONAL);
     // Test that the reservation is deleted
     Assertions.assertThrows(
         NotFoundException.class,
@@ -150,6 +155,9 @@ public class ReservationIT {
     Assertions.assertThrows(
         NotFoundException.class,
         () -> GetReservation.getReservation(PROJECT_ID, RESERVATION_NAME_REGIONAL, ZONE));
+    Assertions.assertThrows(
+        NotFoundException.class,
+        () -> GetReservation.getReservation(PROJECT_ID, RESERVATION_NAME, ZONE));
 
     // Close the client after all tests
     reservationsClient.close();
@@ -170,8 +178,38 @@ public class ReservationIT {
     System.setOut(null);
   }
 
+
   @Test
-  public void testCreateReservationWithGlobalInstanceTemplate()
+  public void firstCreateReservationTest()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+
+    CreateReservation.createReservation(
+        PROJECT_ID, RESERVATION_NAME, NUMBER_OF_VMS, ZONE);
+
+    assertThat(stdOut.toString()).contains("Reservation created. Operation Status: DONE");
+  }
+
+  @Test
+  public void secondGetReservationTest()
+      throws IOException {
+    Reservation reservation = GetReservation.getReservation(
+        PROJECT_ID, RESERVATION_NAME, ZONE);
+
+    assertNotNull(reservation);
+    assertThat(reservation.getName()).isEqualTo(RESERVATION_NAME);
+  }
+
+  @Test
+  public void thirdListReservationTest() throws IOException {
+    List<Reservation> reservations =
+        ListReservations.listReservations(PROJECT_ID, ZONE);
+
+    assertThat(reservations).isNotNull();
+    Assert.assertTrue(reservations.get(0).getName().contains("test-"));
+  }
+
+  @Test
+  public void firstCreateReservationWithGlobalInstanceTemplateTest()
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
     CreateReservationForInstanceTemplate.createReservationForInstanceTemplate(
         PROJECT_ID, RESERVATION_NAME_GLOBAL,
@@ -186,7 +224,7 @@ public class ReservationIT {
   }
 
   @Test
-  public void testCreateReservationWithRegionInstanceTemplate()
+  public void firstCreateReservationWithRegionInstanceTemplateTest()
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
     CreateReservationForInstanceTemplate.createReservationForInstanceTemplate(
         PROJECT_ID, RESERVATION_NAME_REGIONAL, REGIONAL_INSTANCE_TEMPLATE_URI,
